requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
beautifulsoup4==4.12.2
requests==2.31.0
flask==2.3.2
tldextract==3.4.4
ipaddress==1.0.23
dnspython==2.4.2
pyyaml==6.0.1
tqdm==4.65.0
jupyter==1.0.0
notebook==7.0.2
ipykernel==6.25.0
flask-cors==4.0.0
gunicorn==21.2.0
